# plugins/daos_memory/openapi/zeus_memory_action.yaml
openapi: 3.1.0
info:
  title: DAOS Zeus Shared Memory Action
  version: 0.1.0
  description: >-
    Owner-authorized Shared Memory tools for Zeus. Call memory_bootstrap once with a
    one-time Bootstrap Key, retain the returned context_id inside the conversation,
    and pass that context_id to all later operations. Never repeat or expose keys.
servers:
  - url: https://ax.dominof.com
security:
  - BearerAuth: []
paths:
...
  /zeus-memory/v1/events:
    post:
      operationId: memory_write
      summary: Write one Zeus event to Shared Memory
      description: Zeus may write only its registry-authorized memory types and non-protected authority levels. The server fixes source_interface to chatgpt_zeus_action.
      x-openai-isConsequential: true
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WriteRequest'}
      responses:
        '200':
          description: Persisted event with its canonical event ID and source reference.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Event'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
  responses:
...
    BadRequest:
      description: Sanitized invalid request.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Missing/invalid Action API key or expired context handle.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: Zeus writer scope does not permit this event.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    Unavailable:
      description: Bounded sanitized temporary failure.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    ContextId:
      type: string
      minLength: 32
      maxLength: 128
      description: Opaque temporary handle returned by memory_bootstrap. It is useful only with the configured Action API key.
    Product:
      type: string
      minLength: 1
      maxLength: 120
    Topic:
      type: string
      minLength: 1
      maxLength: 160
...
    WriteRequest:
      type: object
      additionalProperties: false
      required: [context_id, product, topic, memory_type, event_type, title, summary, content, authority_level]
      properties:
        context_id: {$ref: '#/components/schemas/ContextId'}
        product: {$ref: '#/components/schemas/Product'}
        topic: {$ref: '#/components/schemas/Topic'}
        memory_type:
          type: string
          enum: [STRATEGY, ASSESSMENT, SESSION_SUMMARY]
        event_type: {type: string, minLength: 1, maxLength: 80}
        title: {type: string, minLength: 1, maxLength: 240}
        summary: {type: string, minLength: 1, maxLength: 1200}
        content: {type: string, minLength: 1, maxLength: 8000}
        authority_level:
          type: string
          enum: [AGENT_ASSESSMENT, HYPOTHESIS, OPERATIONAL_STATE]
        thread_id:
          oneOf: [{type: string, maxLength: 160}, {type: 'null'}]
        work_id:
          oneOf: [{type: string, maxLength: 160}, {type: 'null'}]
        source_ref:
          oneOf: [{type: string, maxLength: 500}, {type: 'null'}]
        metadata: {type: object, additionalProperties: true, default: {}}
...
    Event:
      type: object
      additionalProperties: true
      required: [id]
      properties:
        id: {type: string, format: uuid}
        source_ref:
          oneOf: [{type: string}, {type: 'null'}]
        status: {type: string}
...
    Error:
      type: object
      additionalProperties: false
      required: [ok, error]
      properties:
        ok: {type: boolean, const: false}
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: {type: string, pattern: '^[A-Z][A-Z0-9_]{0,63}$'}
            message: {type: string, maxLength: 200}